mock-gateway/gateway-submitter/requirements.txt
websocket_client
# --extra-index-url http://${HOST_DOCKER_INTERNAL}:${LOCAL_PYPI_PORT}/simple/
# --trusted-host ${HOST_DOCKER_INTERNAL}
# smartgymgatewayutils
requests
redis